+++ internal/upgrade/base.go
@@ -1,0 +1,34 @@
+package upgrade
+
+import (
+	upgradecattlev1 "github.com/rancher/system-upgrade-controller/pkg/apis/upgrade.cattle.io/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+const (
+	upgradeNamespace = "cattle-system"
+)
+
+func baseUpgradePlan(name string) *upgradecattlev1.Plan {
+	const (
+		kind               = "Plan"
+		apiVersion         = "upgrade.cattle.io/v1"
+		serviceAccountName = "system-upgrade-controller"
+	)
+
+	plan := &upgradecattlev1.Plan{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       kind,
+			APIVersion: apiVersion,
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: upgradeNamespace,
+		},
+		Spec: upgradecattlev1.PlanSpec{
+			ServiceAccountName: serviceAccountName,
+		},
+	}
+
+	return plan
+}
